+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * python module: don't emite whitespace in JSON to match serde-json-core (#92)
 
 ### Added
-* `heapless::String` now implements `Miniconf` directly.
+* `heapless::String` and `heapless::Vec` now implement `Miniconf` directly.
 
 ## [0.5.0] - 2022-05-12